// gdr-web-version-mdp/gdr-web-version-mdp/src/app/@data/model/ciclos.ts
export class Ciclos {
  cicloId?: number;
  entidadId?: number;
  titularPuestoId?: number;
  titularPuestoDescripcion?: string;
  jefePuestoId?: number;
  jefePuestoDescripcion?: string;
  gestorPuestoId?: number;
  gestorPuestoDescripcion?: string;
  descripcion?: string;
  anio?: number;
  fechaIni?: string;
  cronogramaId: number;
  fechaFin?: string;
  estadoCicloId?: number;
  estadoCicloDesc?: string;
  estadoRegistro?: string;
  anioString?: string;
}